Fix: 모달 속 댓글 드롭다운 수정 #2
